+++ foxes_opt/objectives/__init__.py
@@ -4,5 +4,6 @@ Wind farm optimization objectives.
 
 from .farm_vars import FarmVarObjective as FarmVarObjective
 from .farm_vars import MaxFarmPower as MaxFarmPower
+from .farm_vars import MaxFarmREWS as MaxFarmREWS
 from .farm_vars import MinimalMaxTI as MinimalMaxTI
 from .max_n_turbines import MaxNTurbines as MaxNTurbines