+++ foxes_opt/pipelines/stages/__init__.py
@@ -1,2 +1,4 @@
 from .ambient_rows import AmbientRowsStage as AmbientRowsStage
 from .greedy_turbine_sweep import GreedyTurbineSweepStage as GreedyTurbineSweepStage
+from .layout_optimizer import LayoutOptimizerStage as LayoutOptimizerStage
+from .random_subset import RandomSubsetStage as RandomSubsetStage